# Cache computed values to reduce build times