--- a/lib/api.ts
+++ b/lib/api.ts
@@ -70,9 +70,15 @@ async function getSlackMessages(slackId: string) {
   }
 }
 
-export async function getContributorsSlugs() {
-  const files = await readdir(`${root}`);
-  return files.map((file) => ({ file }));
+let contributorSlugs: Awaited<ReturnType<typeof getContributorsSlugs>> | null =
+  null;
+
+export async function getContributorsSlugs(): Promise<{ file: string }[]> {
+  if (!contributorSlugs) {
+    const files = await readdir(`${root}`);
+    contributorSlugs = files.map((file) => ({ file }));
+  }
+  return contributorSlugs;
 }
 
 export async function getContributorBySlug(file: string, detail = false) {
@@ -183,11 +189,17 @@ export async function getContributorBySlug(file: string, detail = false) {
   } as Contributor & { summarize: typeof summarize };
 }
 
-export async function getContributors(detail = false) {
-  const slugs = await getContributorsSlugs();
-  return Promise.all(
-    slugs.map((path) => getContributorBySlug(path.file, detail)),
-  );
+let contributors: Awaited<ReturnType<typeof getContributorBySlug>>[] | null =
+  null;
+
+export async function getContributors() {
+  if (!contributors) {
+    const slugs = await getContributorsSlugs();
+    contributors = await Promise.all(
+      slugs.map((path) => getContributorBySlug(path.file)),
+    );
+  }
+  return contributors;
 }
 
 function getCalendarData(activity: Activity[]) {